Implementation of Tools
Completed more work toward the implementation of tools.py to allow for easier use without knowing internals.
Changed the method for handling the relational operator in solvers, now linear solvers will assume a relational of simple equality, if this fails it will fall back on greater than or equal.
Added an import to a levenshtein implementation for some "natural language processing" (https://github.com/rapidfuzz/Levenshtein?tab=readme-ov-file)
Moved around a lot of elements in linearsolvers and renamed it to scipysolvers.
Added vectorization of research and generators for labs w/ research so that costs of science creating factories could be done more accurately.